Refactor EventsController: Split sign-in and home page logic into separate files for improved organization and maintainability. Added detailed comments and documentation for new methods handling event sign-in and home page rendering.
